Fallback to v0.9.0 logic for private gems
Since v0.9.0, the code has been optimized to fetch the release date
directly from the rubygems.org API. Unfortunately this meant that the
release date could no longer be found for gems hosted on private
servers.

In this commit, a `source` attribute is added to the `Gem` model.
Possible values are `:rubygems` and `:other`. The value is decided by
using `::Bundler::LockfileParser` to parse the lockfile and get the
necessary information about each outdated gem.

The `.release_date` method now decides, based on the value of `source`,
whether to call the current implementation, moved into
`.release_date_from_rubygems`, or the implementation from v0.9.0, which
now exists in `.release_date_from_other`.